# public/openapi.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: CI server
servers:
  - url: /api
paths:
  /settings:
    get:
      summary: Return repository settings
      operationId: getSettings
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
    post:
      summary: Saves repository settings
      operationId: saveSettings
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Settings'
        required: true
  /builds:
    get:
      operationId: getBuilds
      summary: Return a list of builds
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Build'
  /builds/{commitHash}:
    post:
      summary: Add specified commit to the build queue
      operationId: runBuild
      parameters:
        - name: commitHash
          in: path
          required: true
          schema:
            type: string
            example: 4ea74703f8294cac67c92e1667fb6b07f539b354
      responses:
        '200':
          description: Success
  /builds/{buildId}:
    get:
      summary: Return build information
      operationId: getBuildInfo
      parameters:
        - name: buildId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Build'
  /builds/{buildId}/logs:
    get:
      summary: Return build logs
      operationId: getBuildLogs
      parameters:
        - name: buildId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: string
components:
  schemas:
    Settings:
      type: object
      required:
        - repoName
        - mainBranch
        - buildCommand
        - period
      properties:
        repoName:
          type: string
          example: some-github-repo
        mainBranch:
          type: string
          example: master
        buildCommand:
          type: string
          example: npm run build
        period:
          type: integer
          description: An interval in minutes between checks for new commits
    Build:
      type: object
      required:
        - buildNumber
        - commitHash
        - commitMessage
        - authorName
        - status
      properties:
        buildNumber:
          type: integer
        commitHash:
          type: string
          example: 4ea74703f8294cac67c92e1667fb6b07f539b354
        commitMessage:
          type: string
          example: improve formatting in readme
        authorName:
          type: string
          example: Dmitriy Davydov
        start:
          type: string
          format: date-time
          example: 2020-01-01T20:15:13Z+05:00
        status:
          type: string
          enum:
            - Waiting
            - InProgress
            - Success
            - Fail
            - Canceled
        duration:
          type: integer
          description: A number of milliseconds that passed while build was running
  links: {}
  callbacks: {}
security: []